+++ src/libxtp/qmpackages/orca.h
@@ -102,13 +102,6 @@ class Orca : public QMPackage {
       {"normal", "Strong"},
       {"tight", "Tight"},
       {"verytight", "VeryTight"}};
-
-  std::unordered_map<std::string, std::string> _libxc_name_to_orca_xc_name{
-      // Hybrid  functionals
-      {"XC_HYB_GGA_XC_B1LYP", "B1LYP"}, // libXC 416
-      {"XC_HYB_GGA_XC_B3LYP", "B3LYP"}, // libXC 402
-      {"XC_HYB_GGA_XC_PBEH", "PBE0"},  // libXC 406
-  };
 };
 
 }  // namespace xtp